Added smart combustion controller

+++ src/main/java/com/bartz24/skyresources/technology/tile/TileRockCrusher.java
@@ -138,7 +138,7 @@ public class TileRockCrusher extends TileGenericPower implements ITickable
 	{
 		super.readFromNBT(compound);
 		bufferListRead(compound.getCompoundTag("buffer"));
-		curProgress = compound.getInteger("progress");
+		curProgress = compound.getFloat("progress");
 	}
 
 	public NBTTagCompound bufferListWrite()
